Create Redundant Braces.java
Given a string of balanced expression. Find whether it contains a redundant parenthesis or not. The redundant parenthesis is the expression surrounded by unnecessary or multiple brackets. Print ‘Yes’ if it's redundant, else ‘No’.



Format:

Input:

The input consists of expression.

Output:

The output consists of the result.



Example:

Input:

((a+b))

Output:

Yes

Input (stdin)
(a+b*(c-d))

Output (stdout)
No